requirements.txt
agno
python-dotenv
qdrant-client
uvicorn[standard] 